Consolidation C-1: proxyclient loses its internal header
First step of the module consolidation (owner decision 2026-07-04:
consolidate now; linting may be disabled selectively where the known
clangd bug produces false errors). Order is consumers-first: a
library's headers can only be deleted once no other package's headers
still include them, so consolidation walks the dependency chain from
the top; the proxy client is the topmost consumer and the smallest.

- LibProxyClientApi.hpp (673 lines) merged into streamrproxyclient.cpp,
  which now imports the four sibling modules (streamr.dht,
  streamr.logger, streamr.trackerlessnetwork, streamr.utils) instead of
  textually including their headers. Third-party libraries, the
  standard library and the public C API header remain textual includes.
- streamr_enable_imports(streamrproxyclient) so the implementation is
  scanned for module dependencies.
- StreamrModules.cmake (canonical + synced copies): the Android NDK
  version floor is now a hard configure-time error instead of a textual
  fall-back - the fall-back required the internal headers that
  consolidation deletes.
- Two NOLINT(bugprone-exception-escape) suppressions: the checker
  cannot see through imported module interfaces (known pattern from the
  import-using test files).
- MODERNIZATION.md: the owner decision, the revised order and the C-1
  record.

Verified: Release build green, proxyclient tests 15/15, standalone
package build green (synthesized module compilation in the consumer
build dir), package lint green.
